+++ docs/more_info/changelog.rst
@@ -4,6 +4,9 @@
 Change Log
 ****************
 
+- Bugfix: Fixed a bug that may cause a crash in case a parameter does not have a minstep value
+- Improvement: DYNAMITE will catch and correct the erroneous parameter generator setting minstep>step by setting minstep=step for non-fixed component parameters
+- Bugfix: Fixed a bug that may occur in the parameter generators (ensures that DYNAMITE creates all possible models)
 - Improvement: now the models of the first two iterations are computed together, better utilizing parallel computing
 - Bugfix: included cmasher in the list of required packages
 - Bugfix: reattempt_failures will no longer result in an error if multiple to-delete models share the same orblib or the orblib directory does not exist